remove all files and use latest code
