+++ src/app/providers.tsx
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import { WalletProviderComponent } from '@/components/solana/WalletProviderComponent';
-import { PaymentProvider } from '@/lib/payment/PaymentContext';
+import { PaymentProvider } from '@/lib/payment/context';
 
 export default function Providers({ children }: { children: React.ReactNode }) {
   return (
